Move codegen parts of WhileDoStmt.cpp to codegen/
