+++ compiler/codegen/Makefile.share
@@ -20,6 +20,7 @@ CODEGEN_SRCS =                                          \
                CForLoop.cpp                             \
                DoWhileStmt.cpp                          \
                LoopStmt.cpp                             \
+               WhileDoStmt.cpp                          \
 
 
 SVN_SRCS = $(CODEGEN_SRCS)
